// src/app/models/issue.ts
export class IssueModel {
    bookId: Number;
    bookName: String;
    memberId: Number;
    memberName: String;
    constructor(){
        
    }
}

// export class ReturnModel {
//     bookId: Number;
//     bookName: String;
//     memberId: Number;
//     memberName: String;
//     issuedate: String;
//     constructor(){
        
//     }
// }